Fixed mismatch with signal/slots from the GUI.

+++ src/modules/ConstellationMgr.cpp
@@ -635,7 +635,7 @@ void ConstellationMgr::update(double deltaTime)
 }
 
 
-void ConstellationMgr::setArtIntensity(float _max)
+void ConstellationMgr::setArtIntensity(double _max)
 {
 	artMaxIntensity = _max;
 	vector < Constellation * >::const_iterator iter;
